Reorganizando react redux y agregando combineReducer
