+++ client/package.json
@@ -25,6 +25,7 @@
     "react-scripts": "4.0.3",
     "react-spinners": "^0.11.0",
     "react-spring": "^9.3.2",
+    "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.4.1",
     "styled-components": "^5.3.3",
