+++ client/src/Components/Pagination/Paginate.jsx
@@ -1,17 +1,17 @@
 import React, { useEffect } from "react";
 import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { getAllProducts } from "../../actions";
+import { getAllProducts } from "../../redux/actions/actionProducts";
 import { ButtonPage, Pagination, LabelPage } from "./stylePag";
 
 export const Paginate = () => {
   const [numberPag, setNumberPag] = useState(1);
-  const totalPages = useSelector((state) => state.totalPages);
-  const category = useSelector((state) => state.category);
+  const totalPages = useSelector((state) => state.products.totalPages);
+  const category = useSelector((state) => state.products.category);
 
   const dispatch = useDispatch();
-  const order = useSelector((state) => state.order);
-  const nameProduct = useSelector((state) => state.nameProduct);
+  const order = useSelector((state) => state.products.order);
+  const nameProduct = useSelector((state) => state.products.nameProduct);
 
   useEffect(() => {
     dispatch(getAllProducts(numberPag, category, order, nameProduct));
