+++ client/src/redux/actions/metodos.js
