+++ client/src/redux/store/index.js
@@ -1,9 +1,10 @@
-import { reducer } from "../reducers/index.js";
 import { createStore, applyMiddleware } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
+import rootRedurcer from "./rootReducer";
+// import { reducer } from "../reducers/productsReducer";
 import thunk from "redux-thunk";
 
 export const store = createStore(
-  reducer,
+  rootRedurcer,
   composeWithDevTools(applyMiddleware(thunk))
 );
