+++ client/src/redux/store/rootReducer.js
@@ -1,0 +1,5 @@
+import { combineReducers } from "redux";
+import { reducer } from "../reducers/productsReducer";
+const rootRedurcer = combineReducers({ products: reducer });
+
+export default rootRedurcer;
